Terminado provider para pagination. Queda la logica del buscador
