+++ src/pages/Pokedex/Pokedex.jsx
@@ -1,6 +1,4 @@
-import {React, useState, useEffect, useContext } from 'react';
-import { getPokemonData } from '../../api';
-
+import {React, useContext } from 'react';
 import PokemonCard from "../../components/PokemonCard/PokemonCard";
 import Loading from '../../components/Loading/Loading';
 import Pagination from '../../components/Pagination/Pagination';
@@ -8,21 +6,8 @@ import { PokemonContext } from '../../contexts/PokemonContext';
 
 const Pokedex = () =>{  
 
-  const {} = useContext(PokemonContext)
+  const {page, total, lastPage, loading, pokemons, nextPage} = useContext(PokemonContext)
 
-    //#PAGINATION
-  useEffect(() => {   
-    fetchPokemons()
-  }, [page])
-  const lastPage =()=> {
-    const nextPage = Math.max(page - 1, 0)
-    setPage(nextPage)
-  }
-  const nextPage =()=> {
-    const nextPage = Math.min(page + 1, total)
-    setPage(nextPage)
-  }
-   
   return(
     <main>
       <div className='coso'>        
@@ -32,9 +17,6 @@ const Pokedex = () =>{
             onLeftClick = {lastPage}
             onRightClick = {nextPage}
           />
-        </div>
-        <div>
-          <h2>hola {num}</h2>
         </div>
         <div>
           {loading?<Loading />
